+++ CHANGES.md
@@ -11,5 +11,6 @@
 
 ### Version 2.0.0
 
-* Add more condition roles and the context level
+* Add more condition roles and the context level as site or course 
 * Creadte seperate table to store easycustmenu settings
+* change the menu control list and menu item form.